Added proc macro for syscall encoding
It's not the prettiest, but it works for now.  A function-like proc
macro might be better, but I'm not sure if there would be an issue
with parsing still because the format I'd like to use is not
fully valid *complete* rust code (hence why I needed to add the empty
{} at the end of each declaration).

+++ filesystems/ext2/src/lib.rs
@@ -11,6 +11,7 @@ use ruxpin_kernel::printkln;
 use ruxpin_kernel::sync::Spinlock;
 use ruxpin_kernel::errors::KernelError;
 
+use ruxpin_kernel::fs::vfs::{is_directory, is_directory_empty};
 use ruxpin_kernel::fs::types::{Filesystem, Mount, Vnode, VnodeOperations, FileAttributes, FilePointer};
 
 mod blocks;
@@ -106,12 +107,16 @@ impl VnodeOperations for Ext2Vnode {
     //}
 
     fn unlink(&mut self, target: Vnode, filename: &str) -> Result<(), KernelError> {
+        if is_directory(target.clone())? && !is_directory_empty(target.clone())? {
+            return Err(KernelError::DirectoryNotEmpty);
+        }
+
         let inode = self.remove_directory_entry(filename)?;
         self.dirty = true;
         self.attrs.nlinks -= 1;
         if self.attrs.nlinks == 0 {
             target.try_lock().unwrap().truncate()?;
-            self.get_mount().superblock.free_inode(inode)?;
+            self.free_inode(inode)?;
         }
         Ok(())
     }
